modularized the code and added the ToString function
